test: ReservationFixture 작성
