+++ app/src/test/java/woowacourse/movie/MovieFixture.kt
@@ -1,5 +1,8 @@
-package woowacourse.movie.domain
+package woowacourse.movie
 
+import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.RunningTime
+import woowacourse.movie.domain.ScreeningDate
 import java.time.LocalDate
 
 val APRIL_THIRTIETH: LocalDate = LocalDate.of(2025, 4, 30)
@@ -21,7 +24,3 @@ val HARRY_POTTER_MOVIE =
             ),
         imageUrl = 0,
     )
-
-val B_CLASS: Point = Point(0, 0)
-val A_CLASS: Point = Point(2, 0)
-val S_CLASS: Point = Point(4, 0)
